Use xed25519_sign instead of curve25519_sign
https://github.com/dino/libomemo-c/commit/c80924f29fceb8d4fb3b9187b4e43d12cacc4136

From the XEP:
-------------
Libraries can use a Curve25519 key pair as their internal IdentityKey.
In this case, the IdentityKey can be used for X25519 directly,
and XEdDSA has to be used to produce EdDSA-compatible signatures.

---

We're using Curve25519 as the internal key pair since this is backwards
compatible with older OMEMO versions.